one part is integrated
